neuron_viewer/requirements.txt
flask==2.3.3
pillow==10.0.0
numpy==1.24.3
torch==2.0.1
transformers==4.31.0
matplotlib==3.7.2
